feat: 댓글 입력 컴포넌트 분리
